metrics moved to cluster state
