+++ cmd/controller/main.go
@@ -44,9 +44,6 @@ import (
 	"github.com/aws/karpenter/pkg/config"
 	"github.com/aws/karpenter/pkg/controllers"
 	"github.com/aws/karpenter/pkg/controllers/counter"
-	metricsnode "github.com/aws/karpenter/pkg/controllers/metrics/node"
-	metricspod "github.com/aws/karpenter/pkg/controllers/metrics/pod"
-	metricsprovisioner "github.com/aws/karpenter/pkg/controllers/metrics/provisioner"
 	"github.com/aws/karpenter/pkg/controllers/node"
 	"github.com/aws/karpenter/pkg/controllers/provisioning"
 	"github.com/aws/karpenter/pkg/controllers/state"
@@ -120,9 +117,9 @@ func main() {
 		state.NewPodController(manager.GetClient(), cluster),
 		node.NewController(manager.GetClient(), cloudProvider, cluster),
 		termination.NewController(ctx, manager.GetClient(), clientSet.CoreV1(), recorder, cloudProvider),
-		metricspod.NewController(manager.GetClient()),
-		metricsnode.NewController(manager.GetClient()),
-		metricsprovisioner.NewController(manager.GetClient()),
+		// metricspod.NewController(manager.GetClient()),
+		// metricsnode.NewController(manager.GetClient()),
+		// metricsprovisioner.NewController(manager.GetClient()),
 		counter.NewController(manager.GetClient(), cluster),
 	).Start(ctx); err != nil {
 		panic(fmt.Sprintf("Unable to start manager, %s", err))
